Breakdown graphql/mutations.py into smaller files
