+++ backend/infrahub/graphql/mutations/__init__.py
@@ -1,0 +1,43 @@
+from .attribute import (
+    AnyAttributeInput,
+    BaseAttributeInput,
+    BoolAttributeInput,
+    CheckboxAttributeInput,
+    ListAttributeInput,
+    NumberAttributeInput,
+    StringAttributeInput,
+    TextAttributeInput,
+)
+from .branch import (
+    BranchCreate,
+    BranchCreateInput,
+    BranchDelete,
+    BranchMerge,
+    BranchNameInput,
+    BranchRebase,
+    BranchValidate,
+)
+from .main import InfrahubMutation, InfrahubMutationMixin, InfrahubMutationOptions
+from .repository import InfrahubRepositoryMutation
+
+__all__ = [
+    "BaseAttributeInput",
+    "BoolAttributeInput",
+    "StringAttributeInput",
+    "NumberAttributeInput",
+    "TextAttributeInput",
+    "CheckboxAttributeInput",
+    "AnyAttributeInput",
+    "ListAttributeInput",
+    "InfrahubRepositoryMutation",
+    "InfrahubMutationOptions",
+    "InfrahubMutation",
+    "InfrahubMutationMixin",
+    "BranchCreate",
+    "BranchCreateInput",
+    "BranchRebase",
+    "BranchValidate",
+    "BranchDelete",
+    "BranchMerge",
+    "BranchNameInput",
+]
